pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "mistune>=3.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]
